requirements.txt
python-telegram-bot
pokepy

